RENAME: (#27) 패키지 이름 오탈자 변경한다
